summarizer/requirements.txt
openai==1.12.0
python-dotenv==1.0.0
pymysql==1.1.0